% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_calcs.R
\name{calculate_factor_stats}
\alias{calculate_factor_stats}
\title{Generate stats for each observation at the factor level}
\usage{
calculate_factor_stats(df, train_data, dep_var, ...)
}
\arguments{
\item{df}{dataframe to evaluate}

\item{train_data}{training dataset generated from summarize_factors}

\item{dep_var}{dependent variable to use (column name)}

\item{...}{Arguments passed on to \code{variation_plot}
\describe{
  \item{df}{dataframe to evaluate}
  \item{dep_var}{dependent variable to use (column name)}
  \item{avg_type}{mean or median}
  \item{ignore_cols}{columns to ignore from analysis. Good candidates are
fields that have have no duplicate values (primary keys) or fields with
a large proportion of null values}
  \item{n_cat}{for categorical variables, the max number of unique values
to keep. This field feeds the \code{forcats::fct_lump(n = )} argument.}
  \item{n_quantile}{for numeric/date fields, the number of quantiles used
to split the data into a factor. Fields that have less than this amount
will not be changed.}
}}
}
\description{
The dataset returned will be the length of the # of columns x # of rows
}
\seealso{
Other manipulation functions: \code{\link{refactor_columns}},
  \code{\link{summarize_factors}}
}
\concept{manipulation functions}
