% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cancer-data.R
\docType{data}
\name{cancer_data}
\alias{cancer_data}
\title{Cancer Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4854 rows and 11 columns.
}
\usage{
cancer_data
}
\description{
This data is a subset of the data used in the \href{https://github.com/cognoma}{Project Cognoma}. Specifically, the samples dataset \href{https://github.com/cognoma/cancer-data/tree/master/data}{here}. Project Cognoma is a Code for Philly and Data Philly collaboration to put "machine learning in the hands of cancer biologists"
}
\keyword{datasets}
