% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{change_range}
\alias{change_range}
\title{Used to expand the range to a new min and max}
\usage{
change_range(x, new_min, new_max)
}
\arguments{
\item{x}{numeric vector}

\item{new_min}{new minimum value}

\item{new_max}{new maxmim value}
}
\description{
Used to expand the range to a new min and max
}
