% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_cut_numeric}
\alias{check_cut_numeric}
\title{confirm if numeric should be cut}
\usage{
check_cut_numeric(x, n_quantile)
}
\arguments{
\item{x}{vector of data}

\item{n_quantile}{(max) number of quantiles to break data}
}
\description{
confirm if numeric should be cut
}
