% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{collapse_cat}
\alias{collapse_cat}
\title{lump categorical data into groups}
\usage{
collapse_cat(x, n)
}
\arguments{
\item{x}{vector of categorical data}

\item{n}{number of categories to keep}
}
\description{
lump categorical data into groups
}
