% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cut_custom}
\alias{cut_custom}
\title{cut numeric and dates into 10 groups}
\usage{
cut_custom(x, n_quantile, order = FALSE)
}
\arguments{
\item{x}{vector of data}

\item{n_quantile}{(max) number of quantiles to break data}

\item{order}{when TRUE, add rank to result ex: "(01) [1,3)", "(02) [4,10]"}
}
\description{
cut numeric and dates into 10 groups
}
