% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{employee_attrition}
\alias{employee_attrition}
\title{Employee Attrition (IBM Watson)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1470 rows and 32 columns.
}
\usage{
employee_attrition
}
\description{
This is a copy of the fictional employee attrition data set created by the Watson team at IBM. The case of the column names have been changed and columns that contained a single value have been dropped. The original data can be found \href{https://github.com/IBM/employee-attrition-aif360}{here} as well as the associated \href{https://www.kaggle.com/pavansubhasht/ibm-hr-analytics-attrition-dataset/home}{Kaggle} competiion. The full data dictionary can be found on the Kaggle link.
}
\keyword{datasets}
