% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_proportions.R
\name{expected_proportions}
\alias{expected_proportions}
\title{Compare representation of proportions (over/under)}
\usage{
expected_proportions(
  df,
  dep_var,
  ...,
  trunc_length = 100,
  sort_by = c("expected", "actual"),
  threshold = 0.02,
  return_data = FALSE,
  n_field = 9,
  color_over = "navyblue",
  color_under = "red"
)
}
\arguments{
\item{df}{data to be analyzed}

\item{dep_var}{dependent variable}

\item{...}{
  Arguments passed on to \code{\link[=variation_plot]{variation_plot}}
  \describe{
    \item{\code{avg_type}}{mean or median}
    \item{\code{ignore_cols}}{columns to ignore from analysis. Good candidates are
fields that have have no duplicate values (primary keys) or fields with
a large proportion of null values}
    \item{\code{n_cat}}{for categorical variables, the max number of unique values
to keep. This field feeds the \code{forcats::fct_lump(n = )} argument.}
    \item{\code{n_quantile}}{for numeric/date fields, the number of quantiles used
to split the data into a factor. Fields that have less than this amount
will not be changed.}
    \item{\code{n_digits}}{for numeric fields, the number of digits to keep in the breaks
ex: [1.2345 to 2.3456] will be [1.23 to 2.34] if \code{n_digits = 2}}
  }}

\item{trunc_length}{length to shorten y-axis labels}

\item{sort_by}{should data be sorted by expected or actual percentages}

\item{threshold}{the cut-off (percentage difference) between actual and
expected values. This allows the chart to focus on the bigger changes.
Use \code{NULL} to keep all values}

\item{return_data}{if TRUE will return a data frame instead of a plot}

\item{n_field}{the max number of facets to show. The fields are sorted in
descending order by those that have the most change (the 'field_delta'
column).}

\item{color_over}{color name/hex code for values that are over-represented}

\item{color_under}{color name/hex code for values that are under-represented}
}
\description{
Compare representation of proportions (over/under)
}
\examples{
expected_proportions(df = iris, dep_var = "Sepal.Length > 5")

# sorted by the expected representation (default)
expected_proportions(
  df = mtcars,
  dep_var = "mpg > 15",
)

# sorted by the actual representation
expected_proportions(
  df = mtcars,
  dep_var = "mpg > 15",
  sort_by = "actual"
)

# you can return the dataframe if you want
expected_proportions(
  df = mtcars,
  dep_var = "mpg > 15",
  return_data = TRUE
)

# an example with more parameters
expected_proportions(
  df = mtcars,          # data to use
  dep_var = "mpg > 15", # can be a field name or an evaluation
  n_cat = 5,            # collapse field values into 5 categories
  n_field = 3,          # keep the frist 3 facets
  threshold = NULL      # keep all values
)
}
