% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{generate_factor_stats}
\alias{generate_factor_stats}
\title{Calculate the importance of the FACTOR in a given field}
\usage{
generate_factor_stats(var, df, avg_fn)
}
\arguments{
\item{var}{variable/field to group by}

\item{df}{refactored data}

\item{avg_fn}{mean or median}
}
\description{
Calculate the importance of the FACTOR in a given field
}
\examples{
# generate_factor_stats("Sepal.Length", refactor_columns(iris, "Sepal.Width"), avg_fn = mean)
}
