% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_calcs.R
\name{generate_field_stats}
\alias{generate_field_stats}
\title{Calculate the importance of the FIELD}
\usage{
generate_field_stats(var, df)
}
\arguments{
\item{var}{variable/field to group by}

\item{df}{refactored data}
}
\description{
Calculate the importance of the FIELD
}
\examples{
generate_field_stats("Sepal.Length", refactor_columns(iris, "Sepal.Width"))
}
