% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-expected-proportions.R
\name{plot_expected_proportions}
\alias{plot_expected_proportions}
\title{Visualize variation between the expected & actual percentages}
\usage{
plot_expected_proportions(
  df,
  dv,
  ...,
  trunc_length = 100,
  sort_by = c("expected", "actual"),
  threshold = 0.02,
  return_data = FALSE,
  n_field = 9,
  color_over = "navyblue",
  color_under = "red"
)
}
\arguments{
\item{df}{data to be analyzed}

\item{dv}{dependent variable}

\item{...}{
  Arguments passed on to \code{\link[=refactor_columns]{refactor_columns}}
  \describe{
    \item{\code{split_on}}{variable to split data / group by}
    \item{\code{id}}{field to use as ID}
    \item{\code{n_cat}}{for categorical variables, the max number of unique values
to keep. This field feeds the \code{forcats::fct_lump(n = )} argument.}
    \item{\code{collapse_by}}{should \code{n_cat} collapse by the distance to the grand
mean \code{"dv"} leaving the extremes as is and grouping factors closer to the
grand mean as "other" or should it use size \code{"n"}}
    \item{\code{n_quantile}}{for numeric/date fields, the number of quantiles used
to split the data into a factor. Fields that have less than this amount
will not be changed.}
    \item{\code{n_digits}}{for numeric fields, the number of digits to keep in the breaks
ex: [1.2345 to 2.3456] will be [1.23 to 2.34] if \code{n_digits = 2}}
    \item{\code{avg_type}}{mean or median}
    \item{\code{ignore_cols}}{columns to ignore from analysis. Good candidates are
fields that have have no duplicate values (primary keys) or fields with
a large proportion of null values}
  }}

\item{trunc_length}{length to shorten y-axis labels}

\item{sort_by}{should data be sorted by expected or actual percentages}

\item{threshold}{the cut-off (percentage difference) between actual and
expected values. This allows the chart to focus on the bigger changes.
Use \code{NULL} to keep all values}

\item{return_data}{if TRUE will return a data frame instead of a plot}

\item{n_field}{the max number of facets to show. The fields are sorted in
descending order by those that have the most change (the 'field_delta'
column).}

\item{color_over}{color name/hex code for values that are over-represented}

\item{color_under}{color name/hex code for values that are under-represented}
}
\description{
Ideal for a 0/1 dichotomous variable.
}
\examples{
plot_expected_proportions(df = iris, dv = Sepal.Length > 5)

# sorted by the expected representation (default)
plot_expected_proportions(
  df = mtcars,
  dv = mpg > 15,
)

# sorted by the actual representation
plot_expected_proportions(
  df = mtcars,
  dv = mpg > 15,
  sort_by = "actual"
)

# you can return the dataframe if you want
plot_expected_proportions(
  df = mtcars,
  dv = mpg > 15,
  return_data = TRUE
)

# an example with more parameters
plot_expected_proportions(
  df = mtcars, # data to use
  dv = mpg > 15, # can be a field name or an evaluation
  n_cat = 5, # collapse field values into 5 categories
  n_field = 3, # keep the frist 3 facets
  threshold = NULL # keep all values
)
}
