% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-group-split.R
\name{plot_group_split}
\alias{plot_group_split}
\title{Visualize variation between two groups}
\usage{
plot_group_split(
  df,
  split_on,
  type = c("dv", "percent", "count"),
  dep_var,
  ...,
  n_cat = 10,
  trunc_length = 100,
  threshold = 0.02,
  base_group = c("1", "2"),
  return_data = FALSE,
  n_field = 9,
  color_over = "navyblue",
  color_under = "red",
  color_missing = "grey50",
  title = NULL,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{df}{dataframe to evaluate}

\item{split_on}{variable to split data / group by}

\item{type}{the outcome or dependent variable ("dv"), the percent of obs.
("percent"), or the number of obs. ("count")}

\item{dep_var}{dependent variable to use (column name)}

\item{...}{
  Arguments passed on to \code{\link[=refactor_columns]{refactor_columns}}
  \describe{
    \item{\code{id}}{field to use as ID}
    \item{\code{collapse_by}}{should \code{n_cat} collapse by the distance to the grand
mean \code{"dv"} leaving the extremes as is and grouping factors closer to the
grand mean as "other" or should it use size \code{"n"}}
    \item{\code{n_quantile}}{for numeric/date fields, the number of quantiles used
to split the data into a factor. Fields that have less than this amount
will not be changed.}
    \item{\code{n_digits}}{for numeric fields, the number of digits to keep in the breaks
ex: [1.2345 to 2.3456] will be [1.23 to 2.34] if \code{n_digits = 2}}
    \item{\code{avg_type}}{mean or median}
    \item{\code{ignore_cols}}{columns to ignore from analysis. Good candidates are
fields that have have no duplicate values (primary keys) or fields with
a large proportion of null values}
  }}

\item{n_cat}{the number of factors to keep in the y-axis. Factors will be
prioritized by the size of the difference and may not match the way
categories are collapsed in \code{refactor_columns()}}

\item{trunc_length}{number of charcters to print on y-axis}

\item{threshold}{threshold for excluding nominal differences. The value
should reflect the type, if the count is in the hundreds you might use 20,
meaning when viewing count differences, values where the difference is <20
will be excluded. For proportion/percent and the dv type, the default is 0.02
or 2 percept}

\item{base_group}{Should group 1 or group 2 be the base. This group will be
the bar and the other will be the point.}

\item{return_data}{When TRUE will return data frame instead of a plot.}

\item{n_field}{How many fields/facets should the plot return.}

\item{color_over}{Color to use when point is higher than bar}

\item{color_under}{Color to use when point is lower than bar}

\item{color_missing}{Color to use when either a point or bar is missing}

\item{title}{title for chart}

\item{subtitle}{subtitle for chart}

\item{caption}{caption for chart}
}
\description{
Visualize variation between two groups
}
\examples{
plot_group_split(ggplot2::mpg, split_on = "year", type = "dv", dep_var = "cty")
plot_group_split(ggplot2::mpg, split_on = "year", type = "dv", dep_var = "cty", base_group = "2")
plot_group_split(ggplot2::mpg, split_on = "year", type = "count", threshold = 10)
plot_group_split(ggplot2::mpg, split_on = "year", type = "percent")
plot_group_split(
  ggplot2::mpg \%>\% dplyr::select(year, cty, trans),
  split_on = "year",
  type = "dv",
  dep_var = "cty",
  base_group = "1", # return_data = TRUE,
  color_missing = "violet"
)

plot_group_split(ggplot2::mpg, split_on = "year", type = "dv", dep_var = "cty", base_group = "1")
plot_group_split(ggplot2::mpg, split_on = "year", type = "dv", dep_var = "cty", base_group = "2")
}
