% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spread.R
\name{plot_spread}
\alias{plot_spread}
\alias{plot_spread_single_obs}
\alias{plot_spread_interactive}
\title{Visualize spread of avg. values among all factors for all variables}
\usage{
plot_spread(df, dv, ..., avg_type = c("mean", "median"))

plot_spread_single_obs(
  df,
  dv,
  ...,
  avg_type = c("mean", "median"),
  labels = FALSE,
  id = 1
)

plot_spread_interactive(...)
}
\arguments{
\item{df}{dataframe to evaluate}

\item{dv}{dependent variable to use (column name)}

\item{...}{
  Arguments passed on to \code{\link[=refactor_columns]{refactor_columns}}, \code{\link[=refactor_columns]{refactor_columns}}, \code{\link[=refactor_columns]{refactor_columns}}
  \describe{
    \item{\code{split_on}}{variable to split data / group by}
    \item{\code{id_col}}{field to use as ID}
    \item{\code{n_cat}}{for categorical variables, the max number of unique values
to keep. This field feeds the \code{forcats::fct_lump(n = )} argument.}
    \item{\code{collapse_by}}{should \code{n_cat} collapse by the distance to the grand
mean \code{"dv"} leaving the extremes as is and grouping factors closer to the
grand mean as "other" or should it use size \code{"n"}}
    \item{\code{n_quantile}}{for numeric/date fields, the number of quantiles used
to split the data into a factor. Fields that have less than this amount
will not be changed.}
    \item{\code{n_digits}}{for numeric fields, the number of digits to keep in the breaks
ex: [1.2345 to 2.3456] will be [1.23 to 2.34] if \code{n_digits = 2}}
    \item{\code{ignore_cols}}{columns to ignore from analysis. Good candidates are
fields that have have no duplicate values (primary keys) or fields with
a large proportion of null values}
  }}

\item{avg_type}{mean or median}

\item{labels}{when TRUE will show the labels of the factor levels outlined in the plot}

\item{id}{is id (row number) from \code{base_data()} to use}
}
\description{
Visualize spread of avg. values among all factors for all variables

Visualize variation and logic for a single observation
}
\section{Functions}{
\itemize{
\item \code{plot_spread_single_obs}: highlight a single observation

\item \code{plot_spread_interactive}: utilizing ggplotly
}}

\examples{
plot_spread(ggplot2::mpg, dv = hwy)
plot_spread_single_obs(df = employee_attrition[,1:5], dv = attrition)
plot_spread_single_obs(df = employee_attrition[,1:5], dv = attrition, labels = TRUE)
plot_spread_interactive(ggplot2::mpg, dv = hwy)
}
