% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_n.R
\name{profile_n}
\alias{profile_n}
\title{Profile examles at top / bottom of estimate}
\usage{
profile_n(df, dep_var, ..., n = 5, position = c("top", "bottom"))
}
\arguments{
\item{df}{dataframe to evaluate}

\item{dep_var}{dependent variable to use (column name)}

\item{...}{
  Arguments passed on to \code{\link[=refactor_columns]{refactor_columns}}
  \describe{
    \item{\code{n_cat}}{for categorical variables, the max number of unique values
to keep. This field feeds the \code{forcats::fct_lump(n = )} argument.}
    \item{\code{n_quantile}}{for numeric/date fields, the number of quantiles used
to split the data into a factor. Fields that have less than this amount
will not be changed.}
    \item{\code{n_digits}}{for numeric fields, the number of digits to keep in the breaks
ex: [1.2345 to 2.3456] will be [1.23 to 2.34] if \code{n_digits = 2}}
    \item{\code{avg_type}}{mean or median}
    \item{\code{ignore_cols}}{columns to ignore from analysis. Good candidates are
fields that have have no duplicate values (primary keys) or fields with
a large proportion of null values}
  }}

\item{n}{number of observations to retrieve}

\item{position}{whether to return the highest ("top") or lowest ("bottom")}
}
\description{
Profile examles at top / bottom of estimate
}
\examples{
\dontrun{
profile_n(df = iris, dep_var = "Sepal.Length")
profile_n(df = iris, dep_var = "Species == 'virginica'")
}

}
\seealso{
Other visualization functions: 
\code{\link{variation_plot_single_obs}()},
\code{\link{variation_plot}()}
}
\concept{visualization functions}
