% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_calcs.R
\name{summarize_factors}
\alias{summarize_factors}
\title{Summarize fields}
\usage{
summarize_factors(..., avg_type = c("mean", "median"))
}
\arguments{
\item{avg_type}{mean or median}
}
\description{
Pivots data and summarizes factor frequencies by field and generates stats
used for plotting
}
