% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-factors.R
\name{summarize_one_field}
\alias{summarize_one_field}
\title{Calculate the importance of the FIELD}
\usage{
summarize_one_field(var, df)
}
\arguments{
\item{var}{variable/field to group by}

\item{df}{refactored data}
}
\description{
Calculate the importance of the FIELD
}
\examples{
# summarize_one_field("Sepal.Length", refactor_columns(iris, "Sepal.Width"))
}
