% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_calcs.R
\name{calculate_factor_stats}
\alias{calculate_factor_stats}
\title{Generate stats for each observation at the factor level}
\usage{
calculate_factor_stats(...)
}
\description{
The dataset returned will be the length of the # of columns x # of rows
}
