% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_drivers.R
\name{find_drivers}
\alias{find_drivers}
\title{Find drivers of variation}
\usage{
find_drivers(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=variation_plot]{variation_plot}}
  \describe{
    \item{\code{df}}{dataframe to evaluate}
    \item{\code{dep_var}}{dependent variable to use (column name)}
    \item{\code{avg_type}}{mean or median}
    \item{\code{ignore_cols}}{columns to ignore from analysis. Good candidates are
fields that have have no duplicate values (primary keys) or fields with
a large proportion of null values}
    \item{\code{n_cat}}{for categorical variables, the max number of unique values
to keep. This field feeds the \code{forcats::fct_lump(n = )} argument.
to keep. This field feeds the \code{forcats::fct_lump(n = )} argument.}
    \item{\code{collapse_by}}{when using n_cat, should the values collapse by the
dependent variable (\code{'dv'}) or size (\code{'n'})}
    \item{\code{n_quantile}}{for numeric/date fields, the number of quantiles used
to split the data into a factor. Fields that have less than this amount
will not be changed.}
    \item{\code{n_digits}}{for numeric fields, the number of digits to keep in the breaks
ex: [1.2345 to 2.3456] will be [1.23 to 2.34] if \code{n_digits = 2}}
  }}
}
\description{
Find drivers of variation
}
\examples{
find_drivers(df = ggplot2::mpg, dep_var = "hwy")
}
