% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_calcs.R
\name{summarize_factors}
\alias{summarize_factors}
\title{Summarize fields}
\usage{
summarize_factors(
  df,
  ...,
  avg_type = c("mean", "median"),
  return = c("data", "list")
)
}
\arguments{
\item{df}{dataframe to evaluate}

\item{...}{
  Arguments passed on to \code{\link[=variation_plot]{variation_plot}}
  \describe{
    \item{\code{dep_var}}{dependent variable to use (column name)}
    \item{\code{ignore_cols}}{columns to ignore from analysis. Good candidates are
fields that have have no duplicate values (primary keys) or fields with
a large proportion of null values}
    \item{\code{n_cat}}{for categorical variables, the max number of unique values
to keep. This field feeds the \code{forcats::fct_lump(n = )} argument.}
    \item{\code{n_quantile}}{for numeric/date fields, the number of quantiles used
to split the data into a factor. Fields that have less than this amount
will not be changed.}
  }}

\item{avg_type}{mean or median}

\item{return}{option to return a dataframe when TRUE and a list when FALSE.
The list option includes the original min/max of the data and the
grand average.}
}
\description{
Pivots data and summarizes factor frequencies by field and generates stats
used for plotting
}
\examples{
summarize_factors(iris, dep_var = "Sepal.Length")
}
\seealso{
Other manipulation functions: 
\code{\link{calculate_factor_stats}()},
\code{\link{refactor_columns}()}
}
\concept{manipulation functions}
