% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_plots.R
\name{variation_plot}
\alias{variation_plot}
\title{Visualize variation among all factor levels for all variables}
\usage{
variation_plot(df, dep_var, n_cat = 10, n_quantile = 10,
  avg_type = c("mean", "median"), ignore_cols = NA_character_)
}
\arguments{
\item{df}{dataframe to evaluate}

\item{dep_var}{dependent variable to use (column name)}

\item{n_cat}{for categorical variables, the max number of unique values
to keep. This field feeds the \code{forcats::fct_lump(n = )} argument.}

\item{n_quantile}{for numeric/date fields, the number of quantiles used
to split the data into a factor. Fields that have less than this amount
will not be changed.}

\item{avg_type}{mean or median}

\item{ignore_cols}{columns to ignore from analysis. Good candidates are
fields that have have no duplicate values (primary keys) or fields with
a large proportion of null values}
}
\description{
Visualize variation among all factor levels for all variables
}
\examples{
variation_plot(ggplot2::mpg, "hwy")
}
\seealso{
Other visualization functions: \code{\link{variation_plot_single_obs}}
}
\concept{visualization functions}
