% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_plots.R
\name{variation_plot_single_obs}
\alias{variation_plot_single_obs}
\title{Visualize variation and logic for a single observation}
\usage{
variation_plot_single_obs(
  df,
  dep_var,
  ...,
  avg_type = c("mean", "median"),
  labels = FALSE,
  id = 1
)
}
\arguments{
\item{df}{dataframe to evaluate}

\item{dep_var}{dependent variable to use (column name)}

\item{...}{
  Arguments passed on to \code{\link[=refactor_columns]{refactor_columns}}
  \describe{
    \item{\code{n_cat}}{for categorical variables, the max number of unique values
to keep. This field feeds the \code{forcats::fct_lump(n = )} argument.}
    \item{\code{n_quantile}}{for numeric/date fields, the number of quantiles used
to split the data into a factor. Fields that have less than this amount
will not be changed.}
    \item{\code{n_digits}}{for numeric fields, the number of digits to keep in the breaks
ex: [1.2345 to 2.3456] will be [1.23 to 2.34] if \code{n_digits = 2}}
    \item{\code{ignore_cols}}{columns to ignore from analysis. Good candidates are
fields that have have no duplicate values (primary keys) or fields with
a large proportion of null values}
  }}

\item{avg_type}{mean or median}

\item{labels}{when TRUE will show the labels of the factor levels outlined in the plot}

\item{id}{is id (row number) from \code{base_data()} to use}
}
\description{
Visualize variation and logic for a single observation
}
\examples{
variation_plot_single_obs(ggplot2::mpg, "hwy")
}
\seealso{
Other visualization functions: 
\code{\link{profile_n}()},
\code{\link{variation_plot}()}
}
\concept{visualization functions}
